+++ sdk/tdf/doc.go
@@ -13,7 +13,7 @@
 // The tdf package enables streaming creation of TDF files with support for:
 //
 //   - Variable-length segments that can arrive out-of-order
-//   - Cryptographic assertions and integrity verification  
+//   - Cryptographic assertions and integrity verification
 //   - Custom attribute-based access controls
 //   - Memory-efficient processing of large datasets
 //   - ZIP archive generation with proper central directory structures
@@ -21,27 +21,27 @@
 // # Basic Usage
 //
 //	ctx := context.Background()
-//	
+//
 //	// Create a new TDF writer
 //	writer, err := tdf.NewWriter(ctx)
 //	if err != nil {
 //		log.Fatal(err)
 //	}
 //	defer writer.Close()
-//	
+//
 //	// Write data segments (can be out-of-order)
 //	data1 := []byte("First segment")
 //	_, err = writer.WriteSegment(ctx, 0, data1)
 //	if err != nil {
 //		log.Fatal(err)
 //	}
-//	
-//	data2 := []byte("Second segment")  
+//
+//	data2 := []byte("Second segment")
 //	_, err = writer.WriteSegment(ctx, 1, data2)
 //	if err != nil {
 //		log.Fatal(err)
 //	}
-//	
+//
 //	// Finalize with attributes and options
 //	finalBytes, manifest, err := writer.Finalize(ctx,
 //		WithAttributeValues(attributes),
@@ -66,8 +66,8 @@
 //
 // The TDF writer uses a two-layer architecture:
 //
-//   1. TDF Layer (tdf.Writer): Handles encryption, assertions, and TDF protocol logic
-//   2. Archive Layer (internal/archive): Manages ZIP file structure and segment assembly
+//  1. TDF Layer (tdf.Writer): Handles encryption, assertions, and TDF protocol logic
+//  2. Archive Layer (internal/archive): Manages ZIP file structure and segment assembly
 //
 // This separation enables independent optimization of cryptographic operations
 // and file format handling.
@@ -88,7 +88,7 @@
 //   - Minimal allocation patterns for high-throughput scenarios
 //
 // Current benchmarks (100 segments, 1KB each):
-//   - Sequential: ~240μs/op, ~530KB memory/op  
+//   - Sequential: ~240μs/op, ~530KB memory/op
 //   - Out-of-order: Similar performance due to contiguous processing optimization
 //
 // # Compatibility
@@ -105,7 +105,7 @@
 // The package uses structured error reporting with operation context:
 //
 //   - ErrAlreadyFinalized: Writer has been finalized
-//   - ErrInvalidSegmentIndex: Invalid segment index provided  
+//   - ErrInvalidSegmentIndex: Invalid segment index provided
 //   - ErrSegmentAlreadyWritten: Duplicate segment index
 //
 // All errors include sufficient context for debugging and recovery.
